Provisioning: Share one Grafana server per package in connection integration tests
Connection integration tests were starting 22 separate Grafana servers — one
per TestIntegration* function. Each startup involves Wire DI, DB provisioning,
HTTP listener, health polling, and shutdown, making these tests very slow.

This refactors the connection package to share a single Grafana server across
all tests using sync.Once + TestMain shutdown:

- Extract StartGrafanaEnvWithManualCleanup in testinfra that returns a cleanup
  func instead of registering on t.Cleanup, so the server can outlive any
  single test.
- Add NewK8sTestHelperShared and RunGrafanaShared that use the manual-cleanup
  variant.
- Rewrite connection/helpers_test.go with sync.Once for lazy init and TestMain
  for shutdown.
- Replace RunGrafana(t) with sharedHelper(t) in all 19 non-enterprise tests
  (one-line change each, preserving top-level TestIntegration* signatures).
- Move 3 enterprise tests to a new connection_enterprise/ package with its own
  shared server and IsEnterprise guard.
- Add CleanupConnections for per-test state isolation.
- Document the strategy in pkg/tests/apis/provisioning/README.md.

Server startups drop from 22 to 2 (one standard, one enterprise).

Made-with: Cursor
