+++ pkg/tests/apis/provisioning/connection/helpers_test.go
@@ -3,9 +3,20 @@ package connection
 import (
 	"testing"
 
-	"github.com/grafana/grafana/pkg/tests/testsuite"
+	ghmock "github.com/migueleliasweb/go-github-mock/src/mock"
+
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 )
 
+var env = common.NewSharedEnv()
+
+func sharedHelper(t *testing.T) *common.ProvisioningTestHelper {
+	t.Helper()
+	helper := env.GetCleanHelper(t)
+	helper.GetEnv().GithubRepoFactory.Client = ghmock.NewMockedHTTPClient()
+	return helper
+}
+
 func TestMain(m *testing.M) {
-	testsuite.Run(m)
+	env.RunTestMain(m)
 }
